LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/sandbox_004-expected.txt



FAIL Fallback content is always displayed for sandboxed PDFs assert_less_than: expected a number less than 300 but got 600
